Move non-test files to a nested folder
